Move weather fetching to fetch_data
