+++ README.md
@@ -4,7 +4,7 @@ This repository contains utilities to download Formula 1 race data using the Jol
 
 ## Data Collection
 
- Raw responses are downloaded with `fetch_data.py` and stored under `jolpica_f1_cache/<season>/<round>.json`. `process_data.py` reads the cache and builds `f1_data_2022_to_present.csv`. Weather data is cached under `weather_cache/` as JSON. A helper script `data_collection.py` runs both steps. The processed data includes:
+ Raw responses and weather information are downloaded with `fetch_data.py` and stored under `jolpica_f1_cache/<season>/<round>.json` and `weather_cache/`. `process_data.py` reads the cached files and builds `f1_data_2022_to_present.csv`. A helper script `data_collection.py` runs both steps. The processed data includes:
 
 - circuit ID
 - start position on the grid
